Subroutines: Problem 1
Assembly program for searching a given integer number in an array of integer numbers. Assume that the numbers in the array are not	in sorted order.
The program must ask the user to enter the number of elements of the array and accept each element of the array through keyboard (for this, use software interrupts).
Also, the user must enter the element to be searched through keyboard. Pass the array and the searching element as parameters	to a subroutine, SEARCH.
The program outputs the position	of the given element, if it is present in the array, otherwise, it outputs	-1.